+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## 1.24.0 (Unreleased)
+
+ENHANCEMENTS:
+
+* resource/ucloud_instance: add `network_interface` to argument in order to create and associate eip with instance.
+
 ## 1.23.0 (2020-09-07)
 
 ENHANCEMENTS: